service/requirements.txt
langchain
langgraph
langchain_openai
openai
fastapi
uvicorn
requests
python-dotenv
python-json-logger
imapclient
python-telegram-bot